+++ src/main/java/org/ateam/oncare/employee/command/repository/EmployeeCareerCommandRepository.java
@@ -3,7 +3,7 @@ package org.ateam.oncare.employee.command.repository;
 import org.ateam.oncare.employee.command.entity.EmployeeCareer;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EmployeeCareerCommandRepository extends JpaRepository<EmployeeCareer, Long> {
+public interface EmployeeCareerCommandRepository extends JpaRepository<EmployeeCareer, Integer> {
     // 직원 수정 시 기존 경력을 싹 지우기 위해 필요
-    void deleteAllByEmployeeId(Long employeeId);
+    void deleteAllByEmployeeId(Integer employeeId);
 }
